Add migration script for quicker project metadata changes
Previously, the template expected the configurations changes to be
manually made by the user. However, there are various ways this could go
wrong and an inexperienced user might not understand which configuration
change is missing from the error messages. This change adds a
quick-start oriented migration script so that one could change the
distributed module name from app.